+++ src/types/mod.rs
@@ -3,6 +3,8 @@ pub mod command;
 pub mod flag;
 pub mod global_flag;
 pub mod load_types;
+pub mod placeholder_definition;
+pub mod snippet;
 pub mod status;
 pub mod template_meta;
 pub mod var_placeholder;